+++ lib/widgets/viewer/controls/controller.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:math';
 
 import 'package:aves/model/entry/entry.dart';
-import 'package:aves/model/entry/extensions/multipage.dart';
 import 'package:aves/services/common/services.dart';
 import 'package:aves/theme/durations.dart';
 import 'package:aves/widgets/viewer/controls/cast.dart';